+++ packages/carbon-graphs/src/js/helpers/axis.js
@@ -881,6 +881,8 @@ const getY2AxisYPosition = (config) => calculateVerticalPadding(config);
 const createAxes = (axis, scale, config, canvasSVG) => {
   getAxesScale(axis, scale, config);
   prepareHAxis(scale, axis, config, prepareHorizontalAxis);
+  // eslint-disable-next-line no-use-before-define
+  const rotateAngle = config.rotateAngle ? calculateRotateAngleBasedOnWidth(config) : undefined;
   canvasSVG
     .append('g')
     .classed(styles.axis, true)
@@ -893,7 +895,11 @@ const createAxes = (axis, scale, config, canvasSVG) => {
       )})`,
     )
     .call(axis.x)
-    .call(resetD3FontSize);
+    .call(resetD3FontSize)
+    .attr('id', 'x')
+    .selectAll('text')
+    .style('text-anchor', config.rotateAngle ? 'end' : 'middle')
+    .attr('transform', () => `rotate(${rotateAngle})`);
   canvasSVG
     .append('g')
     .classed(styles.axis, true)
@@ -922,7 +928,18 @@ const createAxes = (axis, scale, config, canvasSVG) => {
       .call(resetD3FontSize);
   }
 };
-
+const calculateRotateAngleBasedOnWidth = (config) => {
+  const containerWidth = config.canvasWidth;
+  // Define thresholds for rotation based on container width
+  const thresholdSmall = 400;
+  const thresholdMedium = 800;
+  if (containerWidth < thresholdSmall) {
+    return -45; // rotate by -45 degrees for small screens
+  } if (containerWidth < thresholdMedium) {
+    return -30; // rotate by -30 degrees for medium screens
+  }
+  return -15; // rotate by -15 degrees for larger screens
+};
 /**
  * X Axis's position vertically relative to the canvas
  *