+++ packages/carbon-graphs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # ChangeLog
 
 ## Unreleased
+* Fixed
+  * Fixed tick values overlapping on X-axis by rotating 15,30,45.
 
 ## 2.23.3 - (September 25, 2023)